+++ CHANGES.md
@@ -43,6 +43,7 @@
 - macos: always get the current main display and not the main display when the `enigo` struct was created
 - macos: allow dragging all mouse buttons
 - macos: fix crash by executing TIS keyboard layout retrieval on the main thread
+- macos: preserve Shift, Option, and Shift+Option when entering `Key::Unicode(...)` on layout-dependent keys
 - linux: retry with a different protocol when encountering an error
 
 # 0.6.1